requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
python-dotenv==0.19.0
psycopg2-binary==2.9.3
supabase-py==0.0.2
python-jose[cryptography]==3.3.0
requests==2.26.0
gunicorn==20.1.0
